Improved Interchip Communications
Interchip Communications: Now when an Event Monitor thread determines that there is data to send on the communication bus, a new thread (Event Trigger) is created so that the Event Monitor thread can continue processing the other devices triggerable by the event.

Interchip Communications 2: ATMega driver has now been set up with thread locking so that a queue can be formed by multiple Event Trigger threads combined with the already included processing delay so as to not overwhelm the ATMega with more data than it can handle in a given moment.

Audio Recognition + Offset: Added so that events can be triggered before Audio Recognition has identified the movie. The offset allows starting the movie elsewhere other than the beginning.